Refactor feature building (#22)
