+++ .github/workflows/lint.yml
@@ -39,8 +39,9 @@ jobs:
       #----------------------------------------------
       #          run black
       #----------------------------------------------
-      - run: python -m pip install black[jupyter]
+      - run: python -m pip install black==22.12.0
       - run: |
+          black --version
           black ./src ./tests --diff
           black ./src ./tests --check
