+++ pyproject.toml
@@ -26,6 +26,9 @@ priority = "primary"
 [tool.poetry.dependencies]
 python = ">=3.10, <3.11"
 
+google-cloud-bigquery = "^3.26"
+google-cloud-bigquery-storage = "^2.26.0"
+
 # These are RA projects in an internal repo.
 dmlutils = "^0.8.0"
 dmlbootstrap = "^0.2.0"
