+++ requirements-test-14.3.txt
@@ -44,7 +44,7 @@ comm==0.1.2 ; python_version >= "3.10" and python_version < "3.11"
 confection==0.1.4 ; python_version >= "3.10" and python_version < "3.11"
 configparser==5.2.0 ; python_version >= "3.10" and python_version < "3.11"
 contourpy==1.0.5 ; python_version >= "3.10" and python_version < "3.11"
-coverage[toml]==7.6.1 ; python_version >= "3.10" and python_version < "3.11"
+coverage[toml]==7.6.7 ; python_version >= "3.10" and python_version < "3.11"
 cryptography==39.0.1 ; python_version >= "3.10" and python_version < "3.11"
 cycler==0.11.0 ; python_version >= "3.10" and python_version < "3.11"
 cymem==2.0.8 ; python_version >= "3.10" and python_version < "3.11"
@@ -68,9 +68,9 @@ distro==1.9.0 ; python_version >= "3.10" and python_version < "3.11"
 dmlbootstrap==0.2.0 ; python_version >= "3.10" and python_version < "3.11"
 dmlutils==0.8.0 ; python_version >= "3.10" and python_version < "3.11"
 docstring-to-markdown==0.11 ; python_version >= "3.10" and python_version < "3.11"
-duckdb==1.1.1 ; python_version >= "3.10" and python_version < "3.11"
+duckdb==1.1.3 ; python_version >= "3.10" and python_version < "3.11"
 entrypoints==0.4 ; python_version >= "3.10" and python_version < "3.11"
-et-xmlfile==1.1.0 ; python_version >= "3.10" and python_version < "3.11"
+et-xmlfile==2.0.0 ; python_version >= "3.10" and python_version < "3.11"
 evaluate==0.4.1 ; python_version >= "3.10" and python_version < "3.11"
 exceptiongroup==1.2.2 ; python_version >= "3.10" and python_version < "3.11"
 executing==0.8.3 ; python_version >= "3.10" and python_version < "3.11"
@@ -92,8 +92,10 @@ google-api-core==2.15.0 ; python_version >= "3.10" and python_version < "3.11"
 google-api-core[grpc]==2.15.0 ; python_version >= "3.10" and python_version < "3.11"
 google-auth-oauthlib==1.0.0 ; python_version >= "3.10" and python_version < "3.11"
 google-auth==2.21.0 ; python_version >= "3.10" and python_version < "3.11"
+google-cloud-bigquery-storage==2.27.0 ; python_version >= "3.10" and python_version < "3.11"
+google-cloud-bigquery==3.27.0 ; python_version >= "3.10" and python_version < "3.11"
 google-cloud-core==2.4.1 ; python_version >= "3.10" and python_version < "3.11"
-google-cloud-secret-manager==2.20.2 ; python_version >= "3.10" and python_version < "3.11"
+google-cloud-secret-manager==2.21.1 ; python_version >= "3.10" and python_version < "3.11"
 google-cloud-storage==2.11.0 ; python_version >= "3.10" and python_version < "3.11"
 google-crc32c==1.5.0 ; python_version >= "3.10" and python_version < "3.11"
 google-pasta==0.2.0 ; python_version >= "3.10" and python_version < "3.11"
@@ -213,7 +215,7 @@ preshed==3.0.9 ; python_version >= "3.10" and python_version < "3.11"
 prometheus-client==0.14.1 ; python_version >= "3.10" and python_version < "3.11"
 prompt-toolkit==3.0.36 ; python_version >= "3.10" and python_version < "3.11"
 prophet==1.1.5 ; python_version >= "3.10" and python_version < "3.11"
-proto-plus==1.24.0 ; python_version >= "3.10" and python_version < "3.11"
+proto-plus==1.25.0 ; python_version >= "3.10" and python_version < "3.11"
 protobuf==4.24.0 ; python_version >= "3.10" and python_version < "3.11"
 psutil==5.9.0 ; python_version >= "3.10" and python_version < "3.11"
 ptyprocess==0.7.0 ; python_version >= "3.10" and python_version < "3.11"
@@ -249,8 +251,8 @@ pytoolconfig[global]==1.2.5 ; python_version >= "3.10" and python_version < "3.1
 pytz==2022.7 ; python_version >= "3.10" and python_version < "3.11"
 pywavelets==1.4.1 ; python_version >= "3.10" and python_version < "3.11"
 pywin32-ctypes==0.2.3 ; python_version >= "3.10" and python_version < "3.11" and sys_platform == "win32"
-pywin32==307 ; sys_platform == "win32" and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "3.11"
-pywinpty==2.0.13 ; python_version >= "3.10" and python_version < "3.11" and os_name == "nt"
+pywin32==308 ; sys_platform == "win32" and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "3.11"
+pywinpty==2.0.14 ; python_version >= "3.10" and python_version < "3.11" and os_name == "nt"
 pyyaml==6.0 ; python_version >= "3.10" and python_version < "3.11"
 pyzmq==23.2.0 ; python_version >= "3.10" and python_version < "3.11"
 radbcluster==14.3.1 ; python_version >= "3.10" and python_version < "3.11"
