--- src/constants.py
@@ -1,20 +1,0 @@
-# TODO: move into constants.py util
-KILOWATT_HOUR_TO_BRITISH_THERMAL_UNIT = 3412.14
-"""
-The number of kilowatt-hours in British thermal units (BTU)
-"""
-
-BRITISH_THERMAL_UNIT_TO_KILOWATT_HOUR = 1 / KILOWATT_HOUR_TO_BRITISH_THERMAL_UNIT
-"""
-The number British thermal units (BTU) in a kilowatt-hour
-"""
-
-POUND_TO_KILOGRAM = 0.453592
-"""
-The number of pounds in a kilogram.
-"""
-
-KILOGRAM_TO_POUND = 1 / POUND_TO_KILOGRAM
-"""
-The number of kilograms in pound.
-"""
